Create AuthTokenProvider
AuthTokenProvider is an interface with only a function, that returns auth id token, this interface will be only injected to objects related to network requests so backend can validate user.
FirebaseAuthTokenProvider is implementation of AuthTokenProvider, which returns auth id token of a firebase user.
Create provider of AuthTokenProvider so it's ready to be injected as dependency.