Implement note repository.
Note repository's most important job is to, decide on what to do in different connectivity cases.
If there's connectivity, repository sends reequest to backend, if operation finishes succesfully, it only adds Note to NotesCache, and if it fails, it adds Note to both QueueCache and NoteCache.
If there's no connectivity,  it just adds Note to both QueueCache and NoteCache.
Also, getallNotes method only gets notes on local.